重构短信发送功能
